+++ examples/celery_app.rs
@@ -2,7 +2,7 @@
 
 use async_trait::async_trait;
 use celery::error::TaskError;
-use celery::task::{Task, TaskSendOptions};
+use celery::task::Task;
 use env_logger::Env;
 use exitfailure::ExitFailure;
 use structopt::StructOpt;
@@ -72,8 +72,7 @@ async fn main() -> Result<(), ExitFailure> {
             my_app.send_task(add::new(1, 2)).await?;
 
             // Demonstrates sending a task with additional options.
-            let send_options = TaskSendOptions::builder().countdown(10).build();
-            my_app.send_task_with(add::new(1, 3), &send_options).await?;
+            my_app.send_task(add::new(1, 3).with_countdown(10)).await?;
         }
     };