+++ src/app/mod.rs
@@ -2,6 +2,7 @@ use failure::Fail;
 use futures::stream::StreamExt;
 use log::{debug, error, info, warn};
 use std::collections::HashMap;
+use std::convert::TryFrom;
 use tokio::select;
 use tokio::signal::unix::{signal, SignalKind};
 use tokio::sync::mpsc::{self, UnboundedSender};
@@ -14,7 +15,7 @@ mod trace;
 use crate::broker::{Broker, BrokerBuilder};
 use crate::error::{BrokerError, CeleryError, TaskError};
 use crate::protocol::{Message, TryCreateMessage};
-use crate::task::{Signature, Task, TaskEvent, TaskOptions, TaskSendOptions, TaskStatus};
+use crate::task::{Signature, Task, TaskEvent, TaskOptions, TaskStatus};
 use routing::Rule;
 use trace::{build_tracer, TraceBuilder, TracerTrait};
 
@@ -232,33 +233,24 @@ where
         CeleryBuilder::<B::Builder>::new(name, broker_url)
     }
 
-    /// Send a task to a remote worker with default options. Returns the correlation ID
-    /// of the task if successful.
-    pub async fn send_task<T: Task>(&self, task_sig: Signature<T>) -> Result<String, CeleryError> {
-        let queue = routing::route(T::NAME, &self.task_routes).unwrap_or(&self.default_queue);
-        let options = TaskSendOptions::builder().queue(queue).build();
-        self.send_task_with(task_sig, &options).await
-    }
-
-    /// Send a task to a remote worker with custom options. Returns the correlation ID
-    /// of the task if successful.
-    pub async fn send_task_with<T: Task>(
+    /// Send a task to a remote worker. Returns the task ID of the task if successful.
+    pub async fn send_task<T: Task>(
         &self,
-        task_sig: Signature<T>,
-        options: &TaskSendOptions,
+        mut task_sig: Signature<T>,
     ) -> Result<String, CeleryError> {
-        let message = Message::builder(task_sig)?
-            .task_send_options(options)
-            .build();
-        let queue = options.queue.as_ref().unwrap_or(&self.default_queue);
+        let maybe_queue = task_sig.queue.take();
+        let queue = maybe_queue.as_ref().map(|s| s.as_str()).unwrap_or_else(|| {
+            routing::route(T::NAME, &self.task_routes).unwrap_or(&self.default_queue)
+        });
+        let message = Message::try_from(task_sig)?;
         info!(
             "Sending task {}[{}] to {}",
             T::NAME,
-            message.properties.correlation_id,
-            queue
+            message.task_id(),
+            queue,
         );
         self.broker.send(&message, queue).await?;
-        Ok(message.properties.correlation_id)
+        Ok(message.task_id().into())
     }
 
     /// Register a task.