+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- A `Signature` struct.
+- A `Signature` struct with includes task execution options (previously the fields in `TaskSendOptions`).
 - A `bind` argument to the `task` macro. When `bind = true` is given, the task will be run as an instance method.
 - A `Request` struct.
 
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `TaskContext` struct.
+- `TaskSendOptions`.
+- `Celery::send_task_with`.
 
 ## [0.2.0-alpha.2] - 2019-02-24